+++ src/app/features/home/details/actions/actions.page.ts
@@ -1,19 +1,20 @@
 import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute } from '@angular/router';
-import { AlertController } from '@ionic/angular';
-import { AlertInput } from '@ionic/core';
 import { TranslocoService } from '@ngneat/transloco';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { combineLatest } from 'rxjs';
-import { catchError, concatMap, map, tap } from 'rxjs/operators';
+import { catchError, concatMap, first, map, tap } from 'rxjs/operators';
+import { ActionsDialogComponent } from '../../../../shared/actions/actions-dialog/actions-dialog.component';
 import {
   Action,
   ActionsService,
-} from '../../../../shared/actions/actions.service';
+} from '../../../../shared/actions/service/actions.service';
 import { BlockingActionService } from '../../../../shared/blocking-action/blocking-action.service';
 import { DiaBackendAuthService } from '../../../../shared/dia-backend/auth/dia-backend-auth.service';
 import { ErrorService } from '../../../../shared/error/error.service';
+import { isNonNullable } from '../../../../utils/rx-operators/rx-operators';
 
 @UntilDestroy()
 @Component({
@@ -33,12 +34,12 @@ export class ActionsPage {
   constructor(
     private readonly actionsService: ActionsService,
     private readonly errorService: ErrorService,
-    private readonly alertController: AlertController,
     private readonly translocoService: TranslocoService,
     private readonly blockingActionService: BlockingActionService,
     private readonly route: ActivatedRoute,
     private readonly authService: DiaBackendAuthService,
-    private readonly snackBar: MatSnackBar
+    private readonly snackBar: MatSnackBar,
+    private readonly dialog: MatDialog
   ) {}
 
   openAction(action: Action) {
@@ -48,44 +49,25 @@ export class ActionsPage {
       this.id$,
     ])
       .pipe(
-        concatMap(
-          ([params, token, id]) =>
-            new Promise<void>(resolve => {
-              this.alertController
-                .create({
-                  header: action.title_text,
-                  message: action.description_text,
-                  inputs: params.map(
-                    param =>
-                      ({
-                        name: param.name_text,
-                        label: param.display_text_text,
-                        type: param.type_text,
-                        placeholder: param.placeholder_text,
-                        value: param.default_values_list_text[0],
-                        disabled: !param.user_input_boolean,
-                      } as AlertInput)
-                  ),
-                  buttons: [
-                    {
-                      text: this.translocoService.translate('cancel'),
-                      role: 'cancel',
-                    },
-                    {
-                      text: this.translocoService.translate('ok'),
-                      handler: value => {
-                        const body = { ...value, token: token, cid: id };
-                        return this.sendAction(action, body);
-                      },
-                    },
-                  ],
-                })
-                .then(alert => {
-                  alert.present();
-                  resolve();
-                });
-            })
-        ),
+        first(),
+        concatMap(([params, token, id]) => {
+          const dialogRef = this.dialog.open<ActionsDialogComponent>(
+            ActionsDialogComponent,
+            {
+              disableClose: true,
+              data: {
+                action: action,
+                params: params,
+              },
+            }
+          );
+          return dialogRef.afterClosed().pipe(
+            isNonNullable(),
+            tap(data =>
+              this.sendAction(action, { ...data, token: token, cid: id })
+            )
+          );
+        }),
         untilDestroyed(this)
       )
       .subscribe();