+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.44.0 - 2021-12-15
+
+### Changed
+
+- Change network application dialog to support selection input and input validation.
+
+### Fixed
+
+- Do not show Capture options menu when post-creation workflow has not been completed yet
+
 ## 0.43.1 - 2021-12-08
 
 ### Fixed